refactor(study): split StudyRunner god module into focused collaborators
Decompose the ~2,100-line study/runner.py into the orchestrator plus four
sibling modules within the study package (no new infra->study edges):

- study/_progress.py: _QueueProgressCallback + _consume_progress_events
  (pure cross-process progress plumbing, no StudyRunner state).
- study/worker.py: subprocess-worker surface - _run_experiment_worker,
  _collect_result, _derive_exit_reason, _kill_process_group, _UNSET, and the
  COLLECT_RESULT_* constants. Engine/harness imports stay function-local.
- study/baseline_measure.py: _BaselineMixin holding the stateful baseline
  methods verbatim (measurement, caching, drift validation).
- study/image_prep.py: _ImageMixin (_prepare_images, _finalise_image,
  _verify_image_fingerprint) plus pure free functions
  _sanitize_image_for_filename and _parse_image_metadata.
- study/container_lifecycle.py: new free functions persist_failure_artefacts
  and copy_artefact, extracted from the former runner methods.

StudyRunner becomes StudyRunner(_BaselineMixin, _ImageMixin) and keeps the
orchestration surface (__init__, run, _run_one, _run_one_docker,
_handle_result, _save_and_record, etc.). runner.py re-imports the moved worker
and progress symbols so existing import sites and bare-name patch targets keep
resolving. Zero behaviour change; runner.py drops to ~1,015 lines.
